requirements.txt
zabbix-api
python-dotenv
mysql-connector-python
requests